requirements.txt
# Core Flask e WebSocket
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-Cors==5.0.0
Werkzeug==3.1.3

# Eventlet para SocketIO
eventlet==0.39.0
dnspython==2.7.0

# Supabase
supabase==2.10.0

# Impressão térmica
python-escpos==3.1

# Variáveis de ambiente
python-dotenv>=1.0.0

# QR Code
qrcode==8.0

# Processamento de imagens
Pillow==11.0.0

# Notificações Push
pywebpush==1.14.1
cryptography>=42.0.5

# Servidor WSGI para produção (opcional)
gunicorn==23.0.0
